Page: Refactor out section nav state logic and move it to SectionNav component
